% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ae_raw.R
\docType{data}
\name{ae_raw}
\alias{ae_raw}
\title{Adverse Events Raw Dataset}
\format{
A data frame with 32 columns:
\describe{
\item{STUDY}{Study Identifier}
\item{PATNUM}{Patient Number}
\item{FOLDER}{Folder}
\item{FOLDERL}{Folder Label}
\item{IT.AETERM}{Reported Term for the Adverse Event}
\item{AEOUTCOME}{Outcome of Adverse Event}
\item{AELLT}{Lowest Level Term}
\item{AELLTCD}{Lowest Level Term Code}
\item{AEDECOD}{Dictionary-Derived Term}
\item{AEPTCD}{Preferred Term Code}
\item{AEHLT}{High Level Term}
\item{AEHLTCD}{High Level Term Code}
\item{AEHLGT}{High Level Group Term}
\item{AEHLGTCD}{High Level Group Term Code}
\item{AEBODSYS}{Body System or Organ Class}
\item{AEBDSYCD}{Body System or Organ Class Code}
\item{AESOC}{Primary System Organ Class}
\item{AESOCCD}{Primary System Organ Class Code}
\item{IT.AESEV}{Severity/Intensity}
\item{IT.AESER}{Serious Event}
\item{IT.AEREL}{Causality}
\item{IT.AEACN}{Action Taken with Study Treatment}
\item{AESCAN}{Involves Cancer}
\item{AESCNO}{Congenital Anomaly or Birth Defect}
\item{AESDIS}{Persist or Signif Disability/Incapacity}
\item{IT.AESDTH}{Results in Death}
\item{IT.AESHOSP}{Requires or Prolongs Hospitalization}
\item{IT.AESLIFE}{Is Life Threatening}
\item{AESOD}{Occurred with Overdose}
\item{AEDTCOL}{Date/Time of Collection}
\item{IT.AESTDAT}{Start Date/Time of Adverse Event}
\item{IT.AEENDAT}{End Date/Time of Adverse Event}
}
}
\source{
\href{https://github.com/pharmaverse/pharmaverseraw/main/data/ae_raw.rda}{Access the source of the adverse events dataset.}
}
\usage{
ae_raw
}
\description{
A raw dataset used to create pharmaversesdtm::ae
}
\details{
Adverse Events

A raw dataset used to create pharmaversesdtm::ae
}
\author{
Shiyu Chen
}
\keyword{dataset}
