% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_raw.R
\docType{data}
\name{ds_raw}
\alias{ds_raw}
\title{Subject Disposition Raw Dataset}
\format{
A data frame with 13 columns:
\describe{
\item{STUDY}{Study Identifier}
\item{PATNUM}{Patient Number}
\item{SITENM}{Site Name}
\item{INSTANCE}{Instance}
\item{FORM}{Form}
\item{FORML}{Form Label}
\item{IT.DSTERM}{Reason for Discontinuation}
\item{IT.DSDECOD}{Discontinuation Code}
\item{OTHERSP}{Other Reason for Discontinuation}
\item{DSDTCOL}{Date of Collection}
\item{DSTMCOL}{Time of Collection}
\item{IT.DSSTDAT}{Start Date of Subject Disposition}
\item{DEATHDT}{Death Date of Subject}
}
}
\source{
\href{https://github.com/pharmaverse/pharmaverseraw/blob/main/data/ds_raw.rda}{Access the source of the subject disposition dataset.}
}
\usage{
ds_raw
}
\description{
A raw dataset used to create pharmaversesdtm::ds
}
\details{
Subject Disposition

A raw dataset used to create pharmaversesdtm::ds
}
\author{
Shiyu Chen
}
\keyword{dataset}
