% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vs_raw.R
\docType{data}
\name{vs_raw}
\alias{vs_raw}
\title{Vital Signs Raw Dataset}
\format{
A data frame with 15 columns:
\describe{
\item{STUDY}{Study Identifier}
\item{PATNUM}{Patient Number}
\item{INSTANCE}{Folder}
\item{FORM}{Form}
\item{FORML}{Form Label}
\item{VTLD}{Vital Signs Collection Date}
\item{IT.HEIGHT_VSORRES}{Height}
\item{IT.WEIGHT}{Weight}
\item{IT.TEMP}{Temperature}
\item{IT.TEMP_LOC}{Temperature Location}
\item{TMPTC}{Timepoint}
\item{SYS_BP}{Systolic Blood Pressure}
\item{DIA_BP}{Diastolic Blood Pressure}
\item{PULSE}{Pulse}
\item{SUBPOS}{Subject Position}
}
}
\source{
\href{https://github.com/pharmaverse/pharmaverseraw/blob/main/data/vs_raw.rda}{Access the source of the vital signs dataset.}
}
\usage{
vs_raw
}
\description{
A raw dataset used to create pharmaversesdtm::vs
}
\details{
Vital Signs

A raw dataset used to create pharmaversesdtm::vs
}
\author{
Shiyu Chen & Rammprasad Ganapathy
}
\keyword{dataset}
